Correct flights data

Correct flights data to transform IR-measured temperatures into operative temperatures

Usage

correct_flights_data(flights_data, otm_splines)

Arguments

flights_data

A tibble of flights data obtained through the rnp_flights_data function. The tibble must contain columns for `longitude` and `latitude`.

otm_splines

A nested tibble obtained using the gen_otm_splines function. The tibble must contain columns for longitude and latitude and all values must also be in the flights_data

Value

A processed flights_data tibble where IR-measured temperatures (ir_temp) has been corrected to operative temperatures (op_temp).