Elevation profile of the area where we conducted the project at a spatial resolution of ~ 1m^2

Elevation profile of the area where we conducted the project at a spatial resolution of ~ 1m^2

Usage

elevation

Format

A tibble with 8170 rows and 3 columns, all numeric:

longitude

Longitude in decimal degrees

latitude

Latitude in decimal degrees

elevation

Elevation in meters

Examples

elevation
#> # A tibble: 8,170 × 3
#>    latitude longitude elevation
#>       <dbl>     <dbl>     <dbl>
#>  1     39.9     -120.     1243.
#>  2     39.9     -120.     1243.
#>  3     39.9     -120.     1243.
#>  4     39.9     -120.     1244.
#>  5     39.9     -120.     1244.
#>  6     39.9     -120.     1244.
#>  7     39.9     -120.     1244.
#>  8     39.9     -120.     1244.
#>  9     39.9     -120.     1245.
#> 10     39.9     -120.     1245.
#> # ℹ 8,160 more rows