Example flights data

A data set of 33 flights processed through the rnp_flights_data function Each row in the data represents the surface temperature measured recorded in a given tile (i.e., unique latitude and longitude combination) at the moment in time when a flight took place (specified by a unique year, day of the year and minute of the day).

Usage

flights_data

Format

A tibble with 263008 rows and 7 columns, all numeric:

longitude

Longitude of a given tile in decimal degrees

latitude

Latitude of a given tile in decimal degrees

ir_temp

Surface temperature measured via an IR camera

year

Year when the flight took place

doy

Day of the year when the flight took place

mod_start

Minute of the day when the flight started

mod_end

Minute of the day when the flight ended

Examples

flights_data
#> # A tibble: 263,008 × 7
#>    longitude latitude ir_temp  year   doy mod_start mod_end
#>        <dbl>    <dbl>   <dbl> <dbl> <dbl>     <dbl>   <dbl>
#>  1     -120.     39.9    12.5  2023   236       515     519
#>  2     -120.     39.9    14.4  2023   236       515     519
#>  3     -120.     39.9    14.0  2023   236       515     519
#>  4     -120.     39.9    13.7  2023   236       515     519
#>  5     -120.     39.9    14.6  2023   236       515     519
#>  6     -120.     39.9    15.3  2023   236       515     519
#>  7     -120.     39.9    14.5  2023   236       515     519
#>  8     -120.     39.9    14.5  2023   236       515     519
#>  9     -120.     39.9    13.9  2023   236       515     519
#> 10     -120.     39.9    13.9  2023   236       515     519
#> # ℹ 262,998 more rows