Metadata for example flights

Metadata for example flights

Usage

flights_metadata

Format

A data frame with 34 rows and 4 columns:

flight_id

Character column with a unique identifier for each flight

date

Character column with the date when the flight took place in DD/MM/YYYY format

time_start

Character column with the time when the flight started in HH:MM format

time_end

Character column with the time when the flight ended in HH:MM format

Examples

tibble::as_tibble(flights_metadata)
#> # A tibble: 34 × 4
#>    flight_id date      time_start time_end
#>    <chr>     <chr>     <chr>      <chr>   
#>  1 flight01  8/24/2023 8:35       8:39    
#>  2 flight02  8/24/2023 8:59       9:03    
#>  3 flight03  8/24/2023 9:19       9:23    
#>  4 flight04  8/24/2023 9:42       9:46    
#>  5 flight05  8/24/2023 10:06      10:10   
#>  6 flight06  8/24/2023 10:38      10:42   
#>  7 flight07  8/24/2023 10:57      11:01   
#>  8 flight08  8/24/2023 11:20      11:24   
#>  9 flight09  8/24/2023 11:40      11:44   
#> 10 flight10  8/24/2023 12:00      12:04   
#> # ℹ 24 more rows