Example OTM splines nested data

A nested tibble with smoothing splines models for each OTM during each day while deployed in the field to validate the methodology of this package. Each row corresponds to a unique OTM and day of the year combination. The data set contains metadata information on the deployment of the OTMs. This data set was generated from the data set otms_data using the gen_otm_splines function specifying knot_p = 4.

Usage

otms_splines

Format

A tibble of 132 rows and 9 columns:

otm_id

Character column with a unique identifier for each OTM

year

A numeric column indicating year

doy

A numeric column indicating day of the year

microhabitat

Character column describing the microhabitat where the OTM was deployed, between "outrcrop", "small_rock", "rock", "boulder", "ground", and "bush"

orientation

Character column describing the orientation at which the OTM was deployed, between "N", "NE", "E", "SE", "S", "SW", "W", "NW" and "Flat"

latitude

Latitude where the OTM was deployed in decimal degrees

longitude

Longitude where the OTM was deployed in decimal degrees

elevation

Elevation where the OTM was deployed in meters

spline

A nested column containing the smoothing spline model

Examples

otms_splines
#> # A tibble: 132 × 9
#>    otm_id  year   doy microhabitat orientation latitude longitude elevation
#>    <chr>  <dbl> <dbl> <chr>        <chr>          <dbl>     <dbl>     <dbl>
#>  1 OTM01   2023   236 outcrop      N               39.9     -120.     1312.
#>  2 OTM01   2023   237 outcrop      N               39.9     -120.     1312.
#>  3 OTM01   2023   238 outcrop      N               39.9     -120.     1312.
#>  4 OTM01   2023   239 outcrop      N               39.9     -120.     1312.
#>  5 OTM02   2023   236 outcrop      W               39.9     -120.     1313.
#>  6 OTM02   2023   237 outcrop      W               39.9     -120.     1313.
#>  7 OTM02   2023   238 outcrop      W               39.9     -120.     1313.
#>  8 OTM02   2023   239 outcrop      W               39.9     -120.     1313.
#>  9 OTM03   2023   236 rock         W               39.9     -120.     1317.
#> 10 OTM03   2023   237 rock         W               39.9     -120.     1317.
#> # ℹ 122 more rows
#> # ℹ 1 more variable: spline <list>

# when unnesting a spline model
otms_splines$spline[[2]]
#> Call:
#> smooth.spline(x = otm_x$mod, y = otm_x$op_temp, nknots = knot_n)
#> 
#> Smoothing Parameter  spar= 0.1308614  lambda= 5.402194e-08 (15 iterations)
#> Equivalent Degrees of Freedom (Df): 87.04885
#> Penalized Criterion (RSS): 112.3158
#> GCV: 0.201852