Reads and processes data from multiple flights in .tif format and adds metadata

Reads and processes data from multiple flights in .tif format and adds metadata

Usage

rnp_flights_data(path, metadata, resolution)

Arguments

path

A character string indicating the directory where all .tif` files for flights are stored. The name of each flight .tif file must have associated metadata.

metadata

A metadata tibble for the flights with information on the flight_id (character), date (MM/DD/YYYY), time_start (HH:MM) and time_end (HH:MM) of each flight.

resolution

A numeric > 0.5 indicating the spatial resolution of the the final processed output in squared meters. This lower boundary is chosen to exceed the spatial resolution of most drone-mounted thermal imaging cameras.

Value

A data.frame where each row represents a unique tile (a unique combination of x and y UTM coordinates), with an associated surface temperature measurement (surf_temp) for a unique flight. It also provides columns with the year, month month, day of the year doy the minute of the day when the flight started (mod_start) and ended (mod_end)