Matches data with error_max = 100

A tibble indicating the OTM that best describe the dynamics of a given tile assuming a error_max = 100 and a coverage_per = 0.9 (i.e., default settings).The increased error_max is for visualization purposes (i.e., so no empty tiles appear in the plots presented in the documentation vignettes).The flights_data used to perform the matching was corrected using the correct_flights_data function specifying that both a time and a flight-specific temperature correction should be performed.

Usage

matches_100

Format

A tibble of 6386 rows and 4 columns

latitude

A numeric column indicating latitude of a tile

longitude

A numeric column indicating longitude of a tile

otm_id

A character column indicating the OTM that best describe the dynamics of a tile, a value of NA indicates that that tile could not be matched with any OTM with an error < error_max

error

A numeric column indicating the average absolute error between the thermal dynamics of a tile and the OTM that best described it

Examples

matches_100
#> # A tibble: 6,386 × 5
#>    latitude longitude coverage error otm_id
#>       <dbl>     <dbl>    <dbl> <dbl> <chr> 
#>  1     39.9     -120.    0.912  4.38 OTM26 
#>  2     39.9     -120.    0.912  4.55 OTM26 
#>  3     39.9     -120.    0.912  3.48 OTM26 
#>  4     39.9     -120.    0.941  3.26 OTM26 
#>  5     39.9     -120.    0.941  3.55 OTM26 
#>  6     39.9     -120.    0.941  2.85 OTM02 
#>  7     39.9     -120.    0.941  3.31 OTM26 
#>  8     39.9     -120.    0.912  3.11 OTM26 
#>  9     39.9     -120.    0.912  2.96 OTM26 
#> 10     39.9     -120.    0.912  3.66 OTM26 
#> # ℹ 6,376 more rows